Added check to see if the file has already been downloaded and doesn't redownload. Need to add a force flag later

+++ lib/download_product.sh
@@ -9,11 +9,13 @@ download_tile () {
   releases_url="https://network.pivotal.io/api/v2/products/${product}/releases"
   tile_file="$TMPDIR/${product}-${version_token}.pivotal"
 
-  files_url=`curl -qsf -H "Authorization: Token $PIVNET_TOKEN" "$releases_url" | jq --raw-output ".releases[] | select( .version == \"$version\" ) ._links .product_files .href"`
-  download_post_url=`curl -qsf -H "Authorization: Token $PIVNET_TOKEN" $files_url | jq --raw-output ".product_files[] | select( .aws_object_key | endswith(\"pivotal\") ) ._links .download .href"`
-  download_url=`curl -qsf -X POST -d "" -H "Accept: application/json" -H "Content-Type: application/json" -H "Authorization: Token $PIVNET_TOKEN" $download_post_url -w "%{url_effective}\n"`
+  if [ ! -f $tile_file ] ; then
+    files_url=`curl -qsf -H "Authorization: Token $PIVNET_TOKEN" "$releases_url" | jq --raw-output ".releases[] | select( .version == \"$version\" ) ._links .product_files .href"`
+    download_post_url=`curl -qsf -H "Authorization: Token $PIVNET_TOKEN" $files_url | jq --raw-output ".product_files[] | select( .aws_object_key | endswith(\"pivotal\") ) ._links .download .href"`
+    download_url=`curl -qsf -X POST -d "" -H "Accept: application/json" -H "Content-Type: application/json" -H "Authorization: Token $PIVNET_TOKEN" $download_post_url -w "%{url_effective}\n"`
+    curl -qsf -o $tile_file $download_url
+  fi
 
-  curl -qsf -o $tile_file $download_url
   echo $tile_file
 }
 
@@ -25,10 +27,12 @@ download_addon () {
   releases_url="https://network.pivotal.io/api/v2/products/${product}/releases"
   addon_file="$TMPDIR/${product}-${version_token}.tgz"
 
-  files_url=`curl -qsf -H "Authorization: Token $PIVNET_TOKEN" "$releases_url" | jq --raw-output ".releases[] | select( .version == \"$version\" ) ._links .product_files .href"`
-  download_post_url=`curl -qsf -H "Authorization: Token $PIVNET_TOKEN" $files_url | jq --raw-output ".product_files[] | select( .aws_object_key | endswith(\"tgz\") ) ._links .download .href"`
-  download_url=`curl -qsf -X POST -d "" -H "Accept: application/json" -H "Content-Type: application/json" -H "Authorization: Token $PIVNET_TOKEN" $download_post_url -w "%{url_effective}\n"`
+  if [ ! -f $addon_file ] ; then
+    files_url=`curl -qsf -H "Authorization: Token $PIVNET_TOKEN" "$releases_url" | jq --raw-output ".releases[] | select( .version == \"$version\" ) ._links .product_files .href"`
+    download_post_url=`curl -qsf -H "Authorization: Token $PIVNET_TOKEN" $files_url | jq --raw-output ".product_files[] | select( .aws_object_key | endswith(\"tgz\") ) ._links .download .href"`
+    download_url=`curl -qsf -X POST -d "" -H "Accept: application/json" -H "Content-Type: application/json" -H "Authorization: Token $PIVNET_TOKEN" $download_post_url -w "%{url_effective}\n"`
+    curl -qsf -o $tile_file $download_url
+  fi
 
-  curl -qsf -o $tile_file $download_url
   echo $addon_file
 }
